@@ -1,15 +1,15 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion "26.0.2"
+    compileSdkVersion 29
+    buildToolsVersion "26.0.3"
     defaultConfig {
         applicationId "com.wordpress.onelifegroupnz.moaarknatural"
         minSdkVersion 16
-        targetSdkVersion 26
+        targetSdkVersion 29
         versionCode 5
         versionName "1.2"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         release {
@@ -20,23 +20,23 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    androidTestImplementation('androidx.test.espresso:espresso-core:3.1.0', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
 
-    compile 'com.android.support:appcompat-v7:26.1.0'
-    compile 'com.android.support:design:26.1.0'
-    compile 'com.android.support:support-core-utils:26.1.0'
-    compile 'com.dropbox.core:dropbox-core-sdk:3.0.3'
-    compile 'com.google.firebase:firebase-ads:11.6.2'
-    compile 'com.google.firebase:firebase-core:11.6.2'
-    compile 'com.firebase:firebase-jobdispatcher:0.5.2'
-    compile 'com.google.android.gms:play-services-location:11.6.2'
-    compile 'com.google.firebase:firebase-messaging:11.6.2'
-    compile 'com.android.support.constraint:constraint-layout:1.0.2'
-    compile 'com.google.android:flexbox:0.3.0'
-    testCompile 'junit:junit:4.12'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.legacy:legacy-support-core-utils:1.0.0'
+    implementation 'com.dropbox.core:dropbox-core-sdk:3.1.3'
+    implementation 'com.google.firebase:firebase-ads:18.3.0'
+    implementation 'com.google.firebase:firebase-analytics:17.2.1'
+    implementation 'com.firebase:firebase-jobdispatcher:0.5.2'
+    implementation 'com.google.android.gms:play-services-location:17.0.0'
+    implementation 'com.google.firebase:firebase-messaging:20.1.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'com.google.android:flexbox:0.3.0'
+    testImplementation 'junit:junit:4.13'
 }
 
 apply plugin: 'com.google.gms.google-services'
