@@ -2,7 +2,7 @@ package com.wordpress.onelifegroupnz.moaarknatural;
 
 import android.content.Context;
 import android.content.DialogInterface;
-import android.support.v7.app.AlertDialog;
+import androidx.appcompat.app.AlertDialog;
 
 import com.dropbox.core.v2.files.Metadata;
 
